+++ src/hsm/cmd_change_pin.c
@@ -48,11 +48,11 @@ int cmd_change_pin() {
             //encrypt MKEK with new pin
 
             if (P2(apdu) == 0x81) {
-                hash_multi(apdu.data + pin_len, apdu.nc - pin_len, session_pin);
+                hash_multi(apdu.data + pin_len, (uint16_t)(apdu.nc - pin_len), session_pin);
                 has_session_pin = true;
             }
             else if (P2(apdu) == 0x88) {
-                hash_multi(apdu.data + pin_len, apdu.nc - pin_len, session_sopin);
+                hash_multi(apdu.data + pin_len, (uint16_t)(apdu.nc - pin_len), session_sopin);
                 has_session_sopin = true;
             }
             r = store_mkek(mkek);
@@ -62,7 +62,7 @@ int cmd_change_pin() {
             }
             uint8_t dhash[33];
             dhash[0] = (uint8_t)apdu.nc - pin_len;
-            double_hash_pin(apdu.data + pin_len, apdu.nc - pin_len, dhash + 1);
+            double_hash_pin(apdu.data + pin_len, (uint16_t)(apdu.nc - pin_len), dhash + 1);
             flash_write_data_to_file(file_pin, dhash, sizeof(dhash));
             low_flash_available();
             return SW_OK();
